pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["flezzzer <141737861+flezzzer@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
clickhouse-driver = "^0.2.9"
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.37"
clickhouse-sqlalchemy = "^0.3.2"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.10.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
